requirements.txt
pandas==2.3.3
numpy==2.4.0
scikit-learn==1.8.0
xgboost==3.1.2
google-genai==1.56.0
python-dotenv==1.2.1
